[BE] fix: SQL에러 수정
- 없는 값을 참조하려고 해서 발생한 문제

이슈:

+++ back/tink/src/main/java/com/ssafy/tink/dto/dsl/members/FollowInfoDsl.java
@@ -19,12 +19,12 @@ import lombok.NoArgsConstructor;
 @Builder
 public class FollowInfoDsl {
 
-	private int fromId;
+	private Long fromId;
 	@Builder.Default
 	private Set<MemberInfoDsl> member = new HashSet<>();
 
 	@QueryProjection
-	public FollowInfoDsl(int fromId, Set<MemberInfoDsl> member) {
+	public FollowInfoDsl(Long fromId, Set<MemberInfoDsl> member) {
 		this.fromId = fromId;
 		this.member = member;
 	}
